+++ concierge-core/src/main/java/com/grookage/concierge/core/services/ConfigService.java
@@ -1,5 +1,6 @@
 package com.grookage.concierge.core.services;
 
+import com.grookage.concierge.core.cache.ConfigRegistry;
 import com.grookage.concierge.models.SearchRequest;
 import com.grookage.concierge.models.config.ConciergeRequestContext;
 import com.grookage.concierge.models.config.ConfigDetails;
@@ -7,6 +8,7 @@ import com.grookage.concierge.models.config.ConfigKey;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.function.Consumer;
 
 public interface ConfigService {
 
@@ -16,4 +18,5 @@ public interface ConfigService {
 
     List<ConfigDetails> getConfigs(ConciergeRequestContext requestContext, SearchRequest searchRequest);
 
+    Optional<Consumer<ConfigRegistry>> getConfigConsumer(ConciergeRequestContext requestContext);
 }