+++ concierge-core/src/main/java/com/grookage/concierge/core/services/impl/ConfigServiceImpl.java
@@ -1,6 +1,7 @@
 package com.grookage.concierge.core.services.impl;
 
 import com.grookage.concierge.core.cache.CacheConfig;
+import com.grookage.concierge.core.cache.ConfigRegistry;
 import com.grookage.concierge.core.cache.RepositoryRefresher;
 import com.grookage.concierge.core.cache.RepositorySupplier;
 import com.grookage.concierge.core.services.ConfigService;
@@ -10,11 +11,13 @@ import com.grookage.concierge.models.config.ConciergeRequestContext;
 import com.grookage.concierge.models.config.ConfigDetails;
 import com.grookage.concierge.models.config.ConfigKey;
 import com.grookage.concierge.repository.ConciergeRepository;
+import com.grookage.korg.consumer.KorgConsumer;
 import lombok.Builder;
 import lombok.extern.slf4j.Slf4j;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.function.Consumer;
 import java.util.function.Supplier;
 
 @Slf4j
@@ -25,7 +28,9 @@ public class ConfigServiceImpl implements ConfigService {
     private RepositoryRefresher refresher;
 
     @Builder
-    public ConfigServiceImpl(Supplier<ConciergeRepository> repositorySupplier, CacheConfig cacheConfig) {
+    public ConfigServiceImpl(Supplier<ConciergeRepository> repositorySupplier,
+                             CacheConfig cacheConfig,
+                             Supplier<KorgConsumer<ConfigRegistry>> consumerSupplier) {
         this.repositorySupplier = repositorySupplier;
         this.cacheConfig = cacheConfig;
         if (null != cacheConfig && cacheConfig.isEnabled()) {
@@ -35,6 +40,7 @@ public class ConfigServiceImpl implements ConfigService {
                     .supplier(supplier)
                     .dataRefreshInterval(cacheConfig.getRefreshCacheSeconds())
                     .periodicRefresh(true)
+                    .consumerSupplier(consumerSupplier)
                     .build();
             refresher.start();
         }
@@ -62,9 +68,19 @@ public class ConfigServiceImpl implements ConfigService {
                 .toList();
     }
 
+    @Override
+    public Optional<Consumer<ConfigRegistry>> getConfigConsumer(ConciergeRequestContext requestContext) {
+        if (!useRepositoryCache(requestContext)) {
+            log.debug("There is no cache context enabled, returning the empty config consumer");
+            return Optional.empty();
+        }
+        return !useRepositoryCache(requestContext) ? Optional.empty() :
+                Optional.ofNullable(refresher.getConsumerSupplier()).map(Supplier::get);
+    }
+
     private boolean useRepositoryCache(final ConciergeRequestContext requestContext) {
         return null != requestContext && !requestContext.isIgnoreCache()
-                && null != cacheConfig && cacheConfig.isEnabled();
+                && null != cacheConfig && cacheConfig.isEnabled() && null != refresher;
     }
 
     private boolean match(ConfigDetails configDetails, SearchRequest searchRequest) {
@@ -75,6 +91,13 @@ public class ConfigServiceImpl implements ConfigService {
                 searchRequest.getConfigNames().contains(configKey.getConfigName());
         final var configStateMatch = CollectionUtils.isNullOrEmpty(searchRequest.getConfigStates()) ||
                 searchRequest.getConfigStates().contains(configDetails.getConfigState());
-        return namespaceMatch && configNameMatch && configStateMatch;
+        final var orgMatch = CollectionUtils.isNullOrEmpty(searchRequest.getOrgs()) ||
+                searchRequest.getOrgs().contains(configKey.getOrgId());
+        final var tenantMatch = CollectionUtils.isNullOrEmpty(searchRequest.getTenants()) ||
+                searchRequest.getTenants().contains(configKey.getTenantId());
+        final var configTypeMatch = CollectionUtils.isNullOrEmpty(searchRequest.getConfigTypes()) ||
+                searchRequest.getConfigTypes().contains(configKey.getConfigType());
+        return namespaceMatch && configNameMatch && configStateMatch
+                && orgMatch && tenantMatch && configTypeMatch;
     }
 }