+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.0.0]
 
-- Release version for concierge. 
+- Release version for concierge.
 
 ## [0.0.1-RC17]